Move header and footer into 'components' dir
