+++ src/components/footer.svelte
